// frontend/src/api/models/UserAcademicStatusDTO.ts
export interface UserAcademicStatusDTO {
    /**
     * 
     * @type {number}
     * @memberof UserAcademicStatusDTO
     */
    totalRequiredCredit?: number;
    /**
     * 
     * @type {number}
     * @memberof UserAcademicStatusDTO
     */
    totalCompletedCredit?: number;
    /**
     * 
     * @type {number}
     * @memberof UserAcademicStatusDTO
     */
    majorRequiredCredit?: number;
    /**
     * 
     * @type {number}
     * @memberof UserAcademicStatusDTO
     */
    majorCompletedCredit?: number;
    /**
     * 
     * @type {number}
     * @memberof UserAcademicStatusDTO
     */
    majorEssentialRequiredCredit?: number;
    /**
     * 
     * @type {number}
     * @memberof UserAcademicStatusDTO
     */
    majorEssentialCompletedCredit?: number;
    /**
     * 
     * @type {number}
     * @memberof UserAcademicStatusDTO
     */
    liberalRequiredCredit?: number;
    /**
     * 
     * @type {number}
     * @memberof UserAcademicStatusDTO
     */
    liberalCompletedCredit?: number;
    /**
     * 
     * @type {number}
     * @memberof UserAcademicStatusDTO
     */
    liberalEssentialRequiredCredit?: number;
    /**
     * 
     * @type {number}
     * @memberof UserAcademicStatusDTO
     */
    liberalEssentialCompletedCredit?: number;
    /**
     * 
     * @type {number}
     * @memberof UserAcademicStatusDTO
     */
    engineeringRequiredCredit?: number;
    /**
     * 
     * @type {number}
     * @memberof UserAcademicStatusDTO
     */
    engineeringCompletedCredit?: number;
    /**
     * 
     * @type {number}
     * @memberof UserAcademicStatusDTO
     */
    generalRequiredCredit?: number;
    /**
     * 
     * @type {number}
     * @memberof UserAcademicStatusDTO
     */
    generalCompletedCredit?: number;
    /**
     * 
     * @type {number}
     * @memberof UserAcademicStatusDTO
     */
    minimumTotalGradePointAverage?: number;
    /**
     * 
     * @type {number}
     * @memberof UserAcademicStatusDTO
     */
    totalGradePointAverage?: number;
}